I've refactored your Web UI into separate modules to improve its maintainability.
I saw that the main Streamlit UI file, `webui/Main.py`, was over 700 lines long and handled multiple responsibilities, including UI layout, configuration, state management, and business logic. This made the file difficult to read, maintain, and extend.

To improve this, I refactored the UI code by breaking it down into a more organized structure:

- `webui/Main.py`: The main entry point, responsible for the overall page layout and orchestrating calls to UI component functions.
- `webui/ui_components.py`: Contains functions for creating the different parts of the UI (e.g., settings panel, script panel, video panel).
- `webui/utils.py`: Contains utility functions for the web UI (e.g., `get_all_fonts`, `open_task_folder`).
- `webui/i18n.py`: Contains functions for internationalization (`load_locales`, `tr`).

This change improves the code's organization and readability without altering any existing functionality.

Additionally, I made a minor fix by moving a misplaced `from fastapi import Query` import to the top of the file in `app/controllers/v1/video.py` to adhere to standard Python conventions.
